+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/UnicodeTest.php
@@ -5,53 +5,25 @@ declare(strict_types=1);
 namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
 
 use PhpOffice\PhpSpreadsheet\Calculation\Calculation;
-use PhpOffice\PhpSpreadsheet\Calculation\TextData\CharacterConvert as CC;
 use PHPUnit\Framework\Attributes\DataProvider;
 
-class CodeTest extends AllSetupTeardown
+class UnicodeTest extends AllSetupTeardown
 {
-    protected function tearDown(): void
-    {
-        parent::tearDown();
-        CC::setWindowsCharacterSet();
-    }
-
     #[DataProvider('providerCODE')]
     public function testCODE(mixed $expectedResult, mixed $character = 'omitted'): void
     {
         // If expected is array, 1st is for CODE, 2nd for UNICODE,
         // 3rd is for Mac CODE if different from Windows.
         if (is_array($expectedResult)) {
-            $expectedResult = $expectedResult[0];
+            $expectedResult = $expectedResult[1];
         }
         $this->mightHaveException($expectedResult);
         $sheet = $this->getSheet();
         if ($character === 'omitted') {
-            $sheet->getCell('B1')->setValue('=CODE()');
+            $sheet->getCell('B1')->setValue('=UNICODE()');
         } else {
             $this->setCell('A1', $character);
-            $sheet->getCell('B1')->setValue('=CODE(A1)');
-        }
-        $result = $sheet->getCell('B1')->getCalculatedValue();
-        self::assertEquals($expectedResult, $result);
-    }
-
-    #[DataProvider('providerCODE')]
-    public function testMacCODE(mixed $expectedResult, mixed $character = 'omitted'): void
-    {
-        CC::setMacCharacterSet();
-        // If expected is array, 1st is for CODE, 2nd for UNICODE,
-        // 3rd is for Mac CODE if different from Windows.
-        if (is_array($expectedResult)) {
-            $expectedResult = $expectedResult[2] ?? $expectedResult[0];
-        }
-        $this->mightHaveException($expectedResult);
-        $sheet = $this->getSheet();
-        if ($character === 'omitted') {
-            $sheet->getCell('B1')->setValue('=CODE()');
-        } else {
-            $this->setCell('A1', $character);
-            $sheet->getCell('B1')->setValue('=CODE(A1)');
+            $sheet->getCell('B1')->setValue('=UNICODE(A1)');
         }
         $result = $sheet->getCell('B1')->getCalculatedValue();
         self::assertEquals($expectedResult, $result);
@@ -68,7 +40,7 @@ class CodeTest extends AllSetupTeardown
     {
         $calculation = Calculation::getInstance();
 
-        $formula = "=CODE({$array})";
+        $formula = "=UNICODE({$array})";
         $result = $calculation->calculateFormula($formula);
         self::assertSame($expectedResult, $result);
     }